bot/requirements.txt
discord.py>=2.3,<3
asyncssh>=2.14,<3